Move stack to interfaces
