+++ packages/@glimmer/interfaces/index.d.ts
@@ -6,6 +6,7 @@ export * from './lib/module-locators';
 export * from './lib/tier1/symbol-table';
 export * from './lib/template';
 export * from './lib/serialize';
+export * from './lib/stack';
 export { default as ComponentCapabilities } from './lib/component-capabilities';
 
 import * as Simple from './lib/dom/simple';
